reveal controls for patcher selection
